+++ android/src/main/global/io/github/nullptrx/pangleflutter/common/PangleOrientation.kt
@@ -1,7 +1,0 @@
-package io.github.nullptrx.pangleflutter.common
-
-enum class PangleOrientation {
-  none,
-  veritical,
-  horizontal,
-}